+++ frontend/src/Components/Page/Page.js
@@ -67,6 +67,8 @@ class Page extends Component {
       enableColorImpairedMode,
       authenticationEnabled,
       isAdmin,
+      enableDiagnostics,
+      enableDevelopmentMenu,
       onSidebarToggle,
       onSidebarVisibleChange
     } = this.props;
@@ -86,6 +88,8 @@ class Page extends Component {
               isSmallScreen={isSmallScreen}
               isSidebarVisible={isSidebarVisible}
               isAdmin={isAdmin}
+              enableDiagnostics={enableDiagnostics}
+              enableDevelopmentMenu={enableDevelopmentMenu}
               onSidebarVisibleChange={onSidebarVisibleChange}
             />
 
@@ -116,6 +120,8 @@ Page.propTypes = {
   enableColorImpairedMode: PropTypes.bool.isRequired,
   authenticationEnabled: PropTypes.bool.isRequired,
   isAdmin: PropTypes.bool.isRequired,
+  enableDiagnostics: PropTypes.bool.isRequired,
+  enableDevelopmentMenu: PropTypes.bool.isRequired,
   onResize: PropTypes.func.isRequired,
   onSidebarToggle: PropTypes.func.isRequired,
   onSidebarVisibleChange: PropTypes.func.isRequired